requirements.txt
praw>=7.7.0
python-dotenv>=0.21.0
transformers>=4.30.0
torch>=2.0.0
tqdm>=4.0.0